+++ src/Components/App.js
@@ -1,4 +1,4 @@
-import "./App.css";
+import "./Styles/App.css";
 
 function App() {
   return (
