+++ src/Components/Banglore.js
@@ -1,43 +1,42 @@
 import React, { Component } from "react";
+import "./Styles/Banglore.css";
 import { Container } from "react-bootstrap";
 import Row from "react-bootstrap/Row";
 import Col from "react-bootstrap/Col";
-import cloudedsky from "./cloudedsky.jpg";
-
-import "./Home.css";
-class Home extends Component {
+import Bluesky from './Images/Bluesky.jpg';
+class Banglore extends Component {
   render() {
     return (
-      <div classname="white">
+      <div className="break">
         <br></br>
-        <div id="homecss">
+        <div className="home2css">
           <br></br>
           <Container>
             <Row>
               <Col>
                 {" "}
-                <h1 className="head2">My Location</h1>
-              </Col>
-              <Col>
-                {" "}
-                <h2 className="head2">Pithoragarh</h2>
+                <h1 className="head2">Banglore</h1>
               </Col>
 
+              <Col>
+                {" "}
+                <h2 className="head2">2:12 pm</h2>
+              </Col>
               <Row>
                 {" "}
                 <Col>
                   {" "}
-                  <h2 className="head2">cloudy</h2>
+                  <h2 className="head2">Mostly sunny</h2>
                 </Col>
               </Row>
-              <Col>
-                <br></br>
-              </Col>
 
-              <Col>
-                {" "}
-                <h2 className="head2">12°C</h2>
-              </Col>
+              <Row>
+                <Col>
+                  {" "}
+                  <h2 className="head2">26°C</h2>
+                </Col>
+                <br></br>
+              </Row>
             </Row>
           </Container>
         </div>
@@ -46,4 +45,4 @@ class Home extends Component {
   }
 }
 
-export default Home;
+export default Banglore;
