+++ src/Components/Home.js
@@ -1,42 +1,43 @@
 import React, { Component } from "react";
-import "./Banglore.css";
 import { Container } from "react-bootstrap";
 import Row from "react-bootstrap/Row";
 import Col from "react-bootstrap/Col";
-import Bluesky from "./Bluesky.jpg";
-class Banglore extends Component {
+import cloudedsky from "./Images/cloudedsky.jpg";
+
+import "./Styles/Home.css";
+class Home extends Component {
   render() {
     return (
-      <div className="break">
+      <div classname="white">
         <br></br>
-        <div className="home2css">
+        <div id="homecss">
           <br></br>
           <Container>
             <Row>
               <Col>
                 {" "}
-                <h1 className="head2">Banglore</h1>
+                <h1 className="head2">My Location</h1>
+              </Col>
+              <Col>
+                {" "}
+                <h2 className="head2">Pithoragarh</h2>
+              </Col>
+
+              <Row>
+                {" "}
+                <Col>
+                  {" "}
+                  <h2 className="head2">cloudy</h2>
+                </Col>
+              </Row>
+              <Col>
+                <br></br>
               </Col>
 
               <Col>
                 {" "}
-                <h2 className="head2">2:12 pm</h2>
+                <h2 className="head2">12°C</h2>
               </Col>
-              <Row>
-                {" "}
-                <Col>
-                  {" "}
-                  <h2 className="head2">Mostly sunny</h2>
-                </Col>
-              </Row>
-
-              <Row>
-                <Col>
-                  {" "}
-                  <h2 className="head2">26°C</h2>
-                </Col>
-                <br></br>
-              </Row>
             </Row>
           </Container>
         </div>
@@ -45,4 +46,4 @@ class Banglore extends Component {
   }
 }
 
-export default Banglore;
+export default Home;
