+++ src/Components/Styles/Banglore.css
@@ -1,5 +1,5 @@
-.home1css {
-  background-image: url("cloudedsky.jpg");
+.home2css {
+  background-image: url("../Images/Bluesky.jpg");
   height: 200px;
   background-size: cover;
   background-repeat: no-repeat;
