+++ src/Components/Styles/Home.css
@@ -1,5 +1,5 @@
 #homecss{
-    background-image:url("cloudedsky.jpg") ;
+    background-image:url("../Images/cloudedsky.jpg") ;
   height: 200px;
   background-size: cover;
   background-repeat: no-repeat;
