+++ src/Components/Styles/Mylocation.css
@@ -1,5 +1,5 @@
 #img {
-  background-image: url("cloudedsky.jpg");
+  background-image: url("../Images/cloudedsky.jpg");
   height: 100%;
   background-repeat: no-repeat;
   background-size: cover;
